refactor: Rename api services
