+++ src/test/java/no/ssb/dapla/keycloak/mappers/daplauserinfo/DaplaUseInfoMapperIT.java
@@ -1,10 +1,10 @@
 package no.ssb.dapla.keycloak.mappers.daplauserinfo;
 
 import no.ssb.dapla.keycloak.mappers.ConfigPropertyKey;
-import no.ssb.dapla.keycloak.services.teamapi.DaplaApiService;
-import no.ssb.dapla.keycloak.services.teamapi.DaplaTeamApiService;
+import no.ssb.dapla.keycloak.services.api.DaplaApiServiceImpl;
+import no.ssb.dapla.keycloak.services.api.DaplaApiService;
 import no.ssb.dapla.keycloak.services.model.DaplaUserInfo;
-import no.ssb.dapla.keycloak.services.teamapi.DefaultDaplaTeamApiService;
+import no.ssb.dapla.keycloak.services.api.DaplaTeamApiService;
 import no.ssb.dapla.keycloak.utils.Json;
 import org.jboss.logging.Logger;
 import org.junit.jupiter.api.BeforeEach;
@@ -25,7 +25,7 @@ import static no.ssb.dapla.keycloak.Env.requiredEnv;
 @Disabled
 public class DaplaUseInfoMapperIT {
     private static final Logger log = Logger.getLogger(DaplaUserInfoMapperTest.class);
-    private DaplaTeamApiService service;
+    private DaplaApiService service;
 
     private ProtocolMapperModel protocolMapperModel;
     private UserSessionModel userSessionModel;
@@ -37,7 +37,7 @@ public class DaplaUseInfoMapperIT {
 
     @BeforeEach
     void setUp() {
-        service = new DefaultDaplaTeamApiService(DefaultDaplaTeamApiService.Config.builder()
+        service = new DaplaTeamApiService(DaplaTeamApiService.Config.builder()
                 .teamApiUrl(URI.create("http://dapla-team-api.dapla-platform"))
                 .build());
 
@@ -76,7 +76,7 @@ public class DaplaUseInfoMapperIT {
         protocolMapperModel.setConfig(Map.of(
                 ConfigPropertyKey.VERBOSE_LOGGING, Boolean.TRUE.toString(),
                 DaplaUserInfoMapper.ConfigPropertyKey.API_URL, "http://dapla-team-api.dapla-platform",
-                DaplaUserInfoMapper.ConfigPropertyKey.API_IMPL, DefaultDaplaTeamApiService.NAME,
+                DaplaUserInfoMapper.ConfigPropertyKey.API_IMPL, DaplaTeamApiService.NAME,
                 DaplaUserInfoMapper.ConfigPropertyKey.NESTED_TEAMS, Boolean.TRUE.toString(),
                 DaplaUserInfoMapper.ConfigPropertyKey.EXCLUDE_TEAMS_WITHOUT_GROUPS, Boolean.FALSE.toString(),
                 DaplaUserInfoMapper.ConfigPropertyKey.DAPLA_USER_PROPS, "division_code",
@@ -91,7 +91,7 @@ public class DaplaUseInfoMapperIT {
         protocolMapperModel.setConfig(Map.of(
                 ConfigPropertyKey.VERBOSE_LOGGING, Boolean.TRUE.toString(),
                 DaplaUserInfoMapper.ConfigPropertyKey.API_URL, "http://dapla-api.dapla-platform",
-                DaplaUserInfoMapper.ConfigPropertyKey.API_IMPL, DaplaApiService.NAME,
+                DaplaUserInfoMapper.ConfigPropertyKey.API_IMPL, DaplaApiServiceImpl.NAME,
                 DaplaUserInfoMapper.ConfigPropertyKey.NESTED_TEAMS, Boolean.TRUE.toString(),
                 DaplaUserInfoMapper.ConfigPropertyKey.EXCLUDE_TEAMS_WITHOUT_GROUPS, Boolean.FALSE.toString(),
                 DaplaUserInfoMapper.ConfigPropertyKey.DAPLA_USER_PROPS, "division_code",
